REST updated. Locust file updated. Other small updated.

+++ app.py
@@ -2,6 +2,7 @@ from flask import Flask, request, jsonify
 from pymongo import MongoClient
 from bson import json_util, ObjectId
 
+import bson
 import pymongo
 import json
 
@@ -13,6 +14,7 @@ collection = db["cars"]
 
 app = Flask(__name__)
 
+#ADD
 @app.route("/cars", methods=["POST"])
 def add_car():
     car = collection.insert_one({
@@ -27,6 +29,7 @@ def add_car():
     })
     return get_car(str(car.inserted_id))
 
+#GET ALL
 @app.route("/cars", methods=["GET"])
 def get_cars():
     cars = []
@@ -45,9 +48,22 @@ def get_cars():
     
     return jsonify(cars)
 
+#GET BY ID
 @app.route("/cars/<id>", methods=["GET"])
 def get_car(id):
+
+    if not bson.objectid.ObjectId.is_valid(id):
+        return jsonify({
+            'message' : "'id' format is incorrect"
+        })
+
     car = collection.find_one({'_id': ObjectId(id)})
+
+    if car is None:
+        return jsonify({
+            'message' : 'Record does not exist in the database'
+        })
+
     return jsonify({
         '_id': str(ObjectId(car['_id'])),
         'model' : car['model'],
@@ -60,8 +76,22 @@ def get_car(id):
         'gears' : car['gears']
     })
 
+#UPDATE
 @app.route("/cars/<id>", methods=["PUT"])
 def update_car(id):
+
+    #Check if 'id' is correct
+    if not bson.objectid.ObjectId.is_valid(id):
+        return jsonify({
+            'message' : "'id' format is incorrect"
+        })
+
+    #Check if object exists
+    if collection.find_one({'_id': ObjectId(id)}) is None:
+        return jsonify({
+            'message' : 'Record does not exist in the database'
+        })
+    
     collection.update_one({'_id': ObjectId(id)}, {'$set':{
         'model': request.json['model'],
         'color' : request.json['color'],
@@ -72,11 +102,28 @@ def update_car(id):
         'transmission' : request.json['transmission'],
         'gears' : request.json['gears']
     }}) 
-    return get_car(id)
+    return jsonify({
+            'message' : 'Successfully updated'
+        })
 
+#DELETE
 @app.route("/cars/<id>", methods=["DELETE"])
 def delete_car(id):
-    to_delete = get_car(id)
+
+    #Check if 'id' is correct
+    if not bson.objectid.ObjectId.is_valid(id):
+        return jsonify({
+            'message' : "'id' format is incorrect"
+        })
+
+    #Check if object exists
+    if collection.find_one({'_id': ObjectId(id)}) is None:
+        return jsonify({
+            'message' : 'Record does not exist in the database'
+        })
+    
     collection.delete_one({'_id': ObjectId(id)})
-    return to_delete
+    return jsonify({
+            'message' : 'Successfully deleted'
+    })
